# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API pour OC elites du congo",
        "description": "Oc elites du congo",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://elite.com/api"
        }
    ],
    "paths": {
...
        "/admin/stafs/store": {
            "post": {
                "tags": [
                    "ADMINISTRATION__STAFS"
                ],
                "summary": "Ajouter un staf",
                "operationId": "23a9353d0702a4d5a9c0b5db8056cd19",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "postnom": {
                                        "type": "string",
                                        "example": "Kabamba"
                                    },
                                    "date_naissance": {
                                        "type": "date",
                                        "example": "1998-02-18"
                                    },
                                    "lieu_naissance": {
                                        "type": "string",
                                        "example": "Mbuji-mayi"
                                    },
                                    "date_debut_carriere": {
                                        "type": "date",
                                        "example": "2010-08-10"
                                    },
                                    "date_signature": {
                                        "type": "date",
                                        "example": "2020-09-04"
                                    },
                                    "nationalite": {
                                        "type": "string",
                                        "example": "Congolaise"
                                    },
                                    "attribution": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "Une description"
                                    },
                                    "cover": {
                                        "type": "file"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/stafs/show/{id}": {
            "get": {
                "tags": [
                    "ADMINISTRATION__STAFS"
                ],
                "summary": "Renvoie les informations d'un staf par son ID",
                "operationId": "655fc5ae94b447f573d298587883a68b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 18
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/stafs/update": {
            "post": {
                "tags": [
                    "ADMINISTRATION__STAFS"
                ],
                "summary": "Editer les informations d'un staf",
                "operationId": "1a393dc94ede2a56dc30995c00cf5a9b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Mulamba"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Enock"
                                    },
                                    "postnom": {
                                        "type": "string",
                                        "example": "Kabamba"
                                    },
                                    "date_naissance": {
                                        "type": "date",
                                        "example": "1998-02-18"
                                    },
                                    "lieu_naissance": {
                                        "type": "string",
                                        "example": "Mbuji-mayi"
                                    },
                                    "date_debut_carriere": {
                                        "type": "date",
                                        "example": "2010-08-10"
                                    },
                                    "date_signature": {
                                        "type": "date",
                                        "example": "2020-09-04"
                                    },
                                    "nationalite": {
                                        "type": "string",
                                        "example": "Congolaise"
                                    },
                                    "attribution": {
                                        "type": "integer",
                                        "example": 4
                                    },
                                    "descriptions": {
                                        "type": "string",
                                        "example": "Une modif de la description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/stafs/update/image": {
            "post": {
                "tags": [
                    "ADMINISTRATION__STAFS"
                ],
                "summary": "Moofier une photo liée à un évènement",
                "operationId": "125543481dfa7432a0e9ff826ed523f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "staf_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/admin/stafs/delete/image": {
            "post": {
                "tags": [
                    "ADMINISTRATION__STAFS"
                ],
                "summary": "Supprimer une photo liée à un évènement",
                "operationId": "e827009d4e338005c358eaa785309251",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": [
            []
        ]
    }
}